requirements.txt
Django==3.1.7
ipython==7.21.0
psycopg2==2.8.6
django-crispy-forms==1.11.2
djangorestframework==3.12.2
djangorestframework-gis==0.17